+++ tests/ldap/test_utils.py
@@ -78,10 +78,21 @@ def test_filter(slapd_connection, foo_group, bar_group):
     assert Group.query(cn="foo") == [foo_group]
     assert Group.query(cn="bar") == [bar_group]
 
+    assert Group.query(cn="foo") != 3
+
     assert Group.query(cn=["foo"]) == [foo_group]
     assert Group.query(cn=["bar"]) == [bar_group]
 
     assert set(Group.query(cn=["foo", "bar"])) == {foo_group, bar_group}
+
+
+def test_fuzzy(slapd_connection, user, moderator, admin):
+    assert set(User.query()) == {user, moderator, admin}
+    assert set(User.fuzzy("Jack")) == {moderator}
+    assert set(User.fuzzy("moderator")) == {moderator}
+    assert set(User.fuzzy("oderat")) == {moderator}
+    assert set(User.fuzzy("oDeRat")) == {moderator}
+    assert set(User.fuzzy("ack")) == {moderator}
 
 
 def test_ldap_to_python():