+++ CHANGES.rst
@@ -14,12 +14,18 @@ Added
 - A ``populate`` command can be used to fill the database with
   random users generated with faker. :pr:`105`
 - SMTP SSL support. :pr:`108`
+- Server side pagination. :issue:`114` :pr:`111`
 
 Fixed
 *****
 
 - Client deletion also deletes related Consent, Token and
   AuthorizationCode objects. :issue:`126` :pr:`98`
+
+Changed
+*******
+
+- Removed datatables.
 
 [0.0.20] - 2023-01-28
 =====================